+++ src/Data/MediaBus/Conduit/Aggregate.hs
@@ -6,11 +6,17 @@
 module Data.MediaBus.Conduit.Aggregate
   ( aggregateDurationC,
     aggregateCountC,
+    Frames (..),
+    frames,
+    aggregateByC,
+    AggregationResult (..),
   )
 where
 
 import Conduit (ConduitT, await)
+import Control.DeepSeq
 import Control.Lens
+import Data.Foldable (traverse_)
 import Data.MediaBus.Basics.Series (Series (Next, Start))
 import Data.MediaBus.Basics.Ticks
   ( HasDuration (..),
@@ -19,16 +25,129 @@ import Data.MediaBus.Conduit.Stream
   ( yieldNextFrame,
     yieldStartFrameCtx,
   )
+import Data.MediaBus.Media.Channels
+import Data.MediaBus.Media.Media
 import Data.MediaBus.Media.Stream
-  ( Frame (MkFrame),
-    Stream (MkStream),
-    frameCtxSeqNumRef,
-    frameCtxTimestampRef,
-  )
 import Data.Time (NominalDiffTime)
 import Numeric.Natural (Natural)
 
--- | Group content frames into a list of frames such that the total duration
+-- | A list of 'Frame' values aggregated by 'aggregateCountC'
+-- or 'aggregateDurationC'.
+newtype Frames sequenceNumber timestamp payload = MkFrames {_frames :: [Frame sequenceNumber timestamp payload]}
+  deriving (Eq, Ord, NFData, Functor, Monoid, Semigroup)
+
+makeLenses ''Frames
+
+instance
+  (Show s, Show t, HasDuration p) =>
+  Show (Frames s t p)
+  where
+  showsPrec d f@(MkFrames fs) =
+    case fs of
+      [] -> showString "no aggregated frames"
+      (f1 : _) ->
+        showParen
+          (d >= 10)
+          ( showString "frames: starting with "
+              . shows (f1 ^. frameSeqNum)
+              . showString " "
+              . shows (f1 ^. frameTimestamp)
+              . showString " aggregated frames: "
+              . shows (length fs)
+              . showString " total duration: "
+              . shows (getDuration f)
+          )
+
+instance EachFramePayload (Frames s t c) (Frames s t c') where
+  type FramePayloadFrom (Frames s t c) = c
+  type FramePayloadTo (Frames s t c') = c'
+  eachFramePayload = frames . each . eachFramePayload
+
+instance Foldable (Frames sequenceNumber timestamp) where
+  foldr = foldrOf eachFramePayload
+
+instance Traversable (Frames sequenceNumber timestamp) where
+  traverse = eachFramePayload
+
+instance
+  (EachChannel c c') =>
+  EachChannel (Frames s t c) (Frames s t c')
+  where
+  type ChannelsFrom (Frames s t c) = ChannelsFrom c
+  type ChannelsTo (Frames s t c') = ChannelsTo c'
+  eachChannel = eachFramePayload . eachChannel
+
+instance
+  EachMedia c c' =>
+  EachMedia (Frames s t c) (Frames s t c')
+  where
+  type MediaFromE (Frames s t c) = MediaFromE c
+  type MediaToE (Frames s t c') = MediaToE c'
+  eachMedia = eachFramePayload . eachMedia
+
+instance
+  HasDuration c =>
+  HasDuration (Frames s t c)
+  where
+  getDuration = sumOf (eachFramePayload . to getDuration)
+
+-- | The value used to aggregate 'Frames' using 'aggregateByC'.
+data AggregationResult s t c a = MkAggregationResult
+  { aggregationState :: !a,
+    aggregationUnfinished :: !(Frames s t c)
+  }
+
+-- | Group content frames into 'Frames' using a grouping function.
+aggregateByC ::
+  forall a i s t p c m.
+  ( Monad m
+  ) =>
+  -- | Process 'Just' a 'Frame', and the current @AggregationResult@
+  -- return 'Maybe' 'Frames' to yield and an 'AggregationResult'.
+  -- When a start frame is received, or when the conduit ends, the first
+  -- parameter is 'Nothing'.
+  (Maybe (Frame s t c) -> AggregationResult s t c a -> (Maybe (Frames s t c), AggregationResult s t c a)) ->
+  a ->
+  ConduitT
+    (Stream i s t p c)
+    (Stream i () () p (Frames s t c))
+    m
+    ()
+aggregateByC f = go . arInitial
+  where
+    arInitial !initialA =
+      MkAggregationResult
+        { aggregationState = initialA,
+          aggregationUnfinished = mempty
+        }
+
+    go !ar = await >>= maybe handleEOF handlePayload
+      where
+        handleEOF =
+          let (!mOut, _) = f Nothing ar
+           in mapM_ yieldResult mOut
+
+        handlePayload = \case
+          MkStream (Start !pl) -> handleStart pl
+          MkStream (Next !pl) -> handleNext pl
+
+        handleStart !pl = do
+          let (!mRes, !ar') = f Nothing ar
+          traverse_ yieldResult mRes
+          yieldStartFrameCtx
+            ( pl & frameCtxSeqNumRef .~ ()
+                & frameCtxTimestampRef .~ ()
+            )
+          go ar'
+
+        handleNext !pl = do
+          let (!mRes, !ar') = f (Just pl) ar
+          traverse_ yieldResult mRes
+          go ar'
+
+    yieldResult !x = yieldNextFrame (MkFrame () () x)
+
+-- | Group content frames into 'Frames' such that the total duration
 -- of the contents is just @>= t@.
 --
 -- If you want a specific number of contents, and the contents have no 'HasDuration'
@@ -37,52 +156,35 @@ import Numeric.Natural (Natural)
 -- When a start frame is received, the current aggregate is sent,
 -- even if it is incomplete, and the start frame is passed through.
 aggregateDurationC ::
-  forall f i s t p c m.
+  forall i s t p c m.
   ( Monad m,
-    HasDuration c,
-    Applicative f,
-    Semigroup (f (Frame s t c))
+    HasDuration c
   ) =>
   -- | The minimum total duration of the list of payloads.
   NominalDiffTime ->
   ConduitT
     (Stream i s t p c)
-    (Stream i () () p (f (Frame s t c)))
+    (Stream i () () p (Frames s t c))
     m
     ()
-aggregateDurationC t = go Nothing 0
+aggregateDurationC t = aggregateByC f (0 :: NominalDiffTime)
   where
-    go mAcc0 accDur0 =
-      await >>= maybe (yieldAccum mAcc0) handlePayload
-      where
-        handlePayload = \case
-          MkStream (Start pl) -> handleStart pl
-          MkStream (Next pl) -> handleNext pl
-
-        handleStart f = do
-          yieldAccum mAcc0
-          yieldStartFrameCtx
-            ( f & frameCtxSeqNumRef .~ ()
-                & frameCtxTimestampRef .~ ()
-            )
-          go Nothing 0
-
-        handleNext f =
-          let !accDur = accDur0 + getDuration f
-              !acc =
-                maybe
-                  (Just (pure f))
-                  (Just . (<> pure f))
-                  mAcc0
-           in if accDur < t
-                then go acc accDur
-                else do
-                  yieldAccum acc
-                  go Nothing 0
-
-    yieldAccum = \case
-      Nothing -> return ()
-      Just !c -> yieldNextFrame (MkFrame () () c)
+    f Nothing ar =
+      ( if null (aggregationUnfinished ar)
+          then Nothing
+          else Just (aggregationUnfinished ar),
+        ar
+          { aggregationUnfinished = mempty,
+            aggregationState = 0
+          }
+      )
+    f (Just pl) ar =
+      let !accDur = aggregationState ar + getDuration pl
+          !single = MkFrames [pl]
+          !acc = aggregationUnfinished ar <> single
+       in if accDur < t
+            then (Nothing, MkAggregationResult {aggregationState = accDur, aggregationUnfinished = acc})
+            else (Just acc, MkAggregationResult {aggregationState = 0, aggregationUnfinished = mempty})
 
 -- | Group a specific number of content frames into a list of frames.
 --
@@ -92,48 +194,31 @@ aggregateDurationC t = go Nothing 0
 -- When a start frame is received, the current aggregate is sent,
 -- even if it is incomplete, and the start frame is passed through.
 aggregateCountC ::
-  forall f i s t p c m.
-  ( Monad m,
-    Semigroup (f (Frame s t c)),
-    Applicative f
+  forall i s t p c m.
+  ( Monad m
   ) =>
   -- | The minimum total duration of the list of payloads.
   Natural ->
   ConduitT
     (Stream i s t p c)
-    (Stream i () () p (f (Frame s t c)))
+    (Stream i () () p (Frames s t c))
     m
     ()
-aggregateCountC maxCount = go Nothing 0
+aggregateCountC maxCount = aggregateByC f 0
   where
-    go mAcc0 counter0 =
-      await >>= maybe (yieldAccum mAcc0) handlePayload
-      where
-        handlePayload = \case
-          MkStream (Start pl) -> handleStart pl
-          MkStream (Next pl) -> handleNext pl
-
-        handleStart pl = do
-          yieldAccum mAcc0
-          yieldStartFrameCtx
-            ( pl & frameCtxTimestampRef .~ ()
-                & frameCtxSeqNumRef .~ ()
-            )
-          go Nothing 0
-
-        handleNext f =
-          let !counter = counter0 + 1
-              !acc =
-                maybe
-                  (Just (pure f))
-                  (Just . (<> pure f))
-                  mAcc0
-           in if counter < maxCount
-                then go acc counter
-                else do
-                  yieldAccum acc
-                  go Nothing 0
-
-    yieldAccum = \case
-      Nothing -> return ()
-      Just c -> yieldNextFrame (MkFrame () () c)
+    f Nothing ar =
+      ( if null (aggregationUnfinished ar)
+          then Nothing
+          else Just (aggregationUnfinished ar),
+        ar
+          { aggregationUnfinished = mempty,
+            aggregationState = 0
+          }
+      )
+    f (Just pl) ar =
+      let !count = aggregationState ar + 1
+          !single = MkFrames [pl]
+          !acc = aggregationUnfinished ar <> single
+       in if count < maxCount
+            then (Nothing, MkAggregationResult {aggregationState = count, aggregationUnfinished = acc})
+            else (Just acc, MkAggregationResult {aggregationState = 0, aggregationUnfinished = mempty})